Refactor scratch_allocator into its own header
